Write a Go server to serve the HTML content and backend ID validation
So far this isn't a big step up from miniserve, it's just serving the
HTML files but it does validate that the IDs sent to the /students
endpoint match the same regular expression as on the front-end.  Next is
to make it more dynamic and respond with student data instead of just
static content.