Show an error message if the queried student isn't in the list
Originally this was going to give a 404 response but it seems htmx
handles all non-200 codes as irrecoverable errors and doesn't output
their response content.  However, because I want to display the server's
response to the users I've kept it as 200 success, and instead used the
template to output error text instead.  I also added .error styles with
a red background to draw attention to it on the screen because it was a
bit hard to miss in black and white.