Templatify index file to make the ID list dynamic
The main page in this demo lists the valid student IDs you can use to
try out the seearch form.  So far it was static text matching the IDs
I'd written into the actual Go data.  Instead treat this HTML file as a
Go template and print out the IDs dynamically so that they only need to
be modified in one place.  I couldn't find a cool way to print them out
with commas between them in Go templates so I wrote a helper to give
back a list of IDs from the Students data, and that can be joined with
commas using strings.Join.