Add tests for API handlers
In order to test the handlers we need to be able to set them up apart
from when we're starting up the main program.  To do this I copied them
over to a handlers file and add a setupHandlers() function which I call
from both main and from the test setup code.  This way the handlers are
setup exactly the same way in both cases.

The tests so far only focus on the /students endpoint which is what has
the most logic in this app.  It checks that for different kinds of
student requests the API returns the right status code.  This'll come in
handy soon when we change where this data is loaded from; all the tests
should still pass because it's expected that the resulting data and
logic stays the same as it is now.