Read data from external CSV file instead of baked in to the program
In order for someone to use this practically without having to change
the code and recompile it each time the data needs to be somehow read in
externally.  I decided to keep it simple and go with CSV format because
that's easy to parse with the csv lib that's in the standard library,
also easy to edit by non-programmers using spreadsheet apps like Calc or
MS Excel, and it does just what we need for this, no more.