Add test cases for HTTP routing
The setupHandlers function now takes a ServeMux as an argument, so that
for testing we can use a different one than the default.  In future I
may swap this out in main too but for I just keep using the default.

These tests don't test for details of how the students endpoint works,
just to check that it correctly distinguishes between existing and
non-existing paths in the server.