Update to Go 1.22 explicit end of path match for home page
Now instead of checking within the "/" handler whether "/" is the whole
path or not, you can specify a path of "/{$}" which matches only the
root / for the home page and not everything starting with /.  So the
default 404 logic will apply and the home handler only needs to worry
about serving the home page.  Cool beans.